requirements.txt
google-endpoints==2.1.2
google-endpoints-api-management==1.1.3
GoogleAppEngineCloudStorageClient==1.9.22.1
